feat: split types into separate modules, expose proto using _proto suffix
