+++ src/data_types/mod.rs
@@ -1,0 +1,8 @@
+pub mod message;
+pub mod namespace;
+pub mod value;
+pub mod wal;
+
+pub use message::message as message_proto;
+pub use namespace::namespace as namespace_proto;
+pub use wal::wal as wal_proto;
